Move immutable arbor to its own file
