+++ packages/arbor-store/src/Arbor.ts
@@ -1,5 +1,5 @@
 import { DetachedNodeError, NotAnArborNodeError } from "./errors"
-import { hasCustomClonningLogic, isNode } from "./guards"
+import { isNode } from "./guards"
 import { ArrayNodeHandler } from "./handlers/ArrayNodeHandler"
 import { MapNodeHandler } from "./handlers/MapNodeHandler"
 import { NodeHandler } from "./handlers/NodeHandler"
@@ -371,38 +371,5 @@ export class Arbor<T extends object = object> {
    */
   get state(): ArborNode<T> {
     return this.root
-  }
-}
-
-/**
- * @experimental it appears we don't necessarily have to go full immutable to leverage React 18
- * concurrent mode so we'll likely end up removing the 'snapshot' mutation mode idea.
- */
-export class ImmutableArbor<T extends object> extends Arbor<T> {
-  protected readonly engine = new MutationEngine<T>(this, "snapshot")
-
-  cloneNode<V extends object>(node: Node<V>): Node<V> {
-    return this.createNode<V>(
-      this.getPathFor(node),
-      this.clone(node.$value),
-      this.getLinkFor(node),
-      node.$subscriptions
-    )
-  }
-
-  private clone<T extends object>(value: T): T {
-    const clonedValue = hasCustomClonningLogic<T>(value)
-      ? value.$clone()
-      : this.cloneViaConstructor(value)
-
-    const seed = Seed.from(value)
-    Seed.plant(clonedValue, seed)
-    return clonedValue
-  }
-
-  private cloneViaConstructor<T extends object>(value: T): T {
-    const constructor = value.constructor as new () => T
-    const instance = new constructor()
-    return Object.assign(instance, value)
   }
 }
