+++ packages/arbor-store/src/MutationEngine.ts
@@ -21,10 +21,6 @@ export class MutationEngine<T extends object> {
         const childCopy = this.tree.cloneNode(child)
         const link = this.tree.getLinkFor(childCopy)
 
-        // TODO: remove mode, it seems we don't necessarily need to employ
-        // immutability at the data level to get Arbor to work with React's
-        // concurrent mode, as long as we apply structural sharing on the
-        // proxy tree and use `useSyncExternalStore` we should be good.
         if (this.mode === "snapshot") {
           parent.$setChildValue(childCopy.$value, link)
         }
