+++ packages/arbor-store/src/index.ts
@@ -1,10 +1,11 @@
-export { Arbor, ImmutableArbor } from "./Arbor"
+export { Arbor } from "./Arbor"
 export * from "./decorators"
 export * from "./errors"
 export * from "./guards"
 export { ArrayNodeHandler } from "./handlers/ArrayNodeHandler"
 export { MapNodeHandler } from "./handlers/MapNodeHandler"
 export { NodeHandler } from "./handlers/NodeHandler"
+export { ImmutableArbor } from "./ImmutableArbor"
 export * from "./MutationEngine"
 export { Path } from "./path"
 export * from "./ScopedStore"
